// src/features/new-request/info-form/types/types.ts
export type FormState = {
  transportationType: string;
  patientFirstName: string;
  patientLastName: string;
  patientDob: string;
  patientId: string;
  dateOfTransport: string;
  timeOfTransport: string;
  pickupAddress: string;
  destinationAddress: string;
  primaryDiagnosis: string;
  medicalJustification: string;
  formNumber: string;
};